add scoreboard based on save entries

+++ Assets/Scripts/Utils/ScoreEntry.cs
@@ -4,5 +4,5 @@
 public struct ScoreEntry
 {
     public string playerName;
-    public int playTime;
+    public float playTime;
 }
